backend/requirements.txt
Django==3.2.15
django-cors-headers==3.8.0
djangorestframework==3.12.4
django-filter==2.4.0
djoser==2.1.0
django-extra-fields==3.0.2
gunicorn==20.1.0
Pillow==9.0.0
python-dotenv==0.19.0
psycopg2-binary==2.9.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0